+++ build.sbt
@@ -36,11 +36,14 @@ lazy val graalOptions = Seq(
   "-H:+UsePredicates", // SkipFlow optimization -- will be default in GraalVM 25
 )
 
+lazy val TestSerial = config("test-serial") extend Test
+
 lazy val root = (project in file("."))
   .enablePlugins(
     BuildInfoPlugin,
     GraalVMNativeImagePlugin,
   )
+  .configs(TestSerial)
   .settings(
     name := "jelly-cli",
     libraryDependencies ++= Seq(
@@ -52,7 +55,7 @@ lazy val root = (project in file("."))
       ("eu.neverblink.jelly" % "jelly-jena" % jellyV).excludeAll(ExclusionRule("org.apache.jena")),
       "eu.neverblink.jelly" % "jelly-core-protos-google" % jellyV,
       "com.github.alexarchambault" %% "case-app" % "2.1.0",
-      "org.scalatest" %% "scalatest" % "3.2.19" % Test,
+      "org.scalatest" %% "scalatest" % "3.2.19" % "test,test-serial",
       "org.yaml" % "snakeyaml" % "2.4" % Test,
       // For native-image reflection compatibility
       "org.graalvm.sdk" % "graal-sdk" % graalvmV % "provided",
@@ -79,6 +82,11 @@ lazy val root = (project in file("."))
       case PathList("reference.conf") => MergeStrategy.concat
       case _ => MergeStrategy.first
     },
+
+    // Serial tests should not run in parallel.
+    // They are used for tests that manipulate global state, like system properties.
+    inConfig(TestSerial)(Defaults.testSettings),
+    TestSerial / parallelExecution := false,
 
     // GraalVM settings
     Compile / mainClass := Some("eu.neverblink.jelly.cli.App"),
